asserts moved to public file
